Add files via upload
This is a rough draft of the website. Wasn't sure what we wanted for color scheme so that can be changed. Also wasn't sure what our gym name was so need to add that. I am also planning on adding a javascript file so that there can be a form that new members can fill out. Let me know if you see anything else